feat: latest; going off for a bit

+++ astromine-core/src/main/java/com/github/mixinors/astromine/common/block/entity/FluidCollectorBlockEntity.java
@@ -64,7 +64,7 @@ public class FluidCollectorBlockEntity extends ComponentEnergyFluidBlockEntity i
 	}
 
 	@Override
-	public double getEnergyConsumed() {
+	public long getEnergyConsumed() {
 		return AMConfig.get().fluidCollectorEnergyConsumed;
 	}
