+++ astromine-core/src/main/java/com/github/mixinors/astromine/common/block/entity/machine/EnergyConsumedProvider.java
@@ -25,5 +25,5 @@
 package com.github.mixinors.astromine.common.block.entity.machine;
 
 public interface EnergyConsumedProvider {
-	double getEnergyConsumed();
+	long getEnergyConsumed();
 }
